Factor out simple LineIntIndex
